feat: Allow generate plugins
