+++ snarkjs-generate-solidity/templates/verifier_fflonk.sol.ejs
