+++ snarkjs-generate-solidity/templates/verifier_groth16.sol.ejs
