+++ snarkjs-generate-solidity/templates/verifier_plonk.sol.ejs
